QuantumTransport.jl

Documentation for QuantumTransport.jl.

QuantumTransport._helperIncludeModules — Method

_helperIncludeModules(dir)

Function to include modules in all subdirectories of the given directory where all module files are titled Module.jl.

Arguments

dir: The directory to include modules from.

QuantumTransport._importAndExportModules — Function

_importAndExportModules()

Dynamically imports and exports modules within the QuantumTransport module.

This function iterates over the modules defined within the QuantumTransport module and dynamically imports them using the using statement. It then exports all the names defined within each module using the export statement.

No arguments are required for this function.

QuantumTransport._includeModulesInSubdirs — Method

_includeModulesInSubdirs()

This function is responsible for calling _helperIncludeModules(item), where item is each top-level subdirectory of /src. QuantumTransport module.

QuantumTransport.WoodburyModule.block_inversion — Function

block_inversion(matrix, threshold=8)

Compute the inverse of a block matrix using a recursive algorithm.

Arguments

matrix: The block matrix to be inverted.

threshold: The maximum size of the blocks at which the algorithm switches to directly computing the inverse using inv().

Returns

The inverse of the input block matrix.

QuantumTransport.WoodburyModule.generate_matrix — Method

generate_matrix()

Generate a matrix based on some criteria.

QuantumTransport.InputOutputModule.get_data — Method

get_data(readDir::String, filename::String)

Reads data from a file. The first row of the csv file should contain information about the type of plot. The second row of the file should contain the column headings. The third row onwards should contain the data.

Arguments

readDir::String: The directory path where the file is located.

filename::String: The name of the file to read.

Returns

The data read from the file.