API

QuantumTransport.InputOutputModule.get_data — Function

get_data(readDir::String, filename::String)

Reads data from a file. The first row of the csv file should contain information about the type of plot. The second row of the file should contain the column headings. The third row onwards should contain the data.

Arguments

readDir::String: The directory path where the file is located.

filename::String: The name of the file to read.

Returns

The data read from the file.

QuantumTransport.WoodburyModule.block_inversion — Function

block_inversion(matrix, threshold=8)

Compute the inverse of a block matrix using a recursive algorithm.

Arguments

matrix: The block matrix to be inverted.

threshold: The maximum size of the blocks at which the algorithm switches to directly computing the inverse using inv().

Returns

The inverse of the input block matrix.

QuantumTransport.WoodburyModule.generate_matrix — Function

generate_matrix()

Generate a matrix based on some criteria.